findPointsFromStart and End functions modified
What changes you have made?
previously didn't check whether the distance in a valid strcture -
'positive number'

Why this change is necessary?
if the distance is not a positive number greater than zero it will crash

What is the impact of the change?
if given a distance other than expected format it will return null