hasPoint and covers function of circle modified
What changes you have made?
it was not checking whether the given point is an instance of Point
class or not

Why this change is necessary?
It should expect only point of a Point class object

What is the impact of the change?
if the point is not an instance of point it will return false