isEqualTo function of rectangle modified
What changes you have made?
it was not checking whether the given rectangle is an instance of Rectangle
class or not

Why this change is necessary?
It should expect only Rectangle of a Rectangle class object

What is the impact of the change?
if the Rectangle is not an instance of Rectangle it will return false